Import devel/hdf5-c++ version 1.8.7.
HDF5 is a data model, library, and file format for storing and
managing data. It supports an unlimited variety of datatypes, and
is designed for flexible and efficient I/O and for high volume and
complex data. HDF5 is portable and is extensible, allowing applications
to evolve in their use of HDF5. The HDF5 Technology suite includes
tools and applications for managing, manipulating, viewing, and
analyzing data in the HDF5 format.

This package provides C++ wrappers for the HDF5 C library.