Import pkgtools/tinderbox-dragonfly
Tinderbox is a well-known package building system for FreeBSD.
This commit ports the original Tinderbox to DragonFly.  The scripts
are highly platform specific, so only one operating system can be
supported per package.  With interest, alternative versions such as
tinderbox-netbsd and tinderbox-sunos could be developed fairly easily
as they would share the pkgsrc-specific changes with this package.

The repository for the project is:
https://github.com/jrmarino/pkgsrc-tinderbox