devel/tet3: Multiple platform fix
Tet3 has a korn shell script in it, but this was never accounted for
previously.  NetBSD has /bin/ksh so they never noticed the error.  However,
the script in question didn't have the shebang at the first line because
it was a template and REPLACE_KSH failed.  As a result, the template is
patched to allow REPLACE_KSH to work.  This fix should apply to all
platforms that don't have /usr/ksh.