Update php53 pacakge to 5.3.9; PHP 5.3.9. suhosin-patch is provided as modified one; only copyright year.
PHP 5.3.9 Released!

[10-Jan-2012] The PHP development team would like to announce the immediate
availability of PHP 5.3.9. This release focuses on improving the stability of
the PHP 5.3.x branch with over 90 bug fixes, some of which are security
related.

Security Enhancements and Fixes in PHP 5.3.9:

	* Added max_input_vars directive to prevent attacks based on hash
	  collisions. (CVE-2011-4885)
	* Fixed bug #60150 (Integer overflow during the parsing of invalid
          exif header). (CVE-2011-4566)

Key enhancements in PHP 5.3.9 include:

	* Fixed bug #55475 (is_a() triggers autoloader, new optional 3rd
          argument to is_a and is_subclass_of).
	* Fixed bug #55609 (mysqlnd cannot be built shared)
	* Many changes to the FPM SAPI module

For a full list of changes in PHP 5.3.9, see the ChangeLog. For source
downloads please visit our downloads page, Windows binaries can be found on
windows.php.net/download/.

All users are strongly encouraged to upgrade to PHP 5.3.9.