Importing package for CPAN distribution Module::Implementation version 0.06 into devel/p5-Module-Distribution as requirement for scheduled update of devel/p5-Class-Load to 0.18.
This module abstracts out the process of choosing one of several
underlying implementations for a module. This can be used to provide
XS and pure Perl implementations of a module, or it could be used to
load an implementation for a given OS or any other case of needing
to provide multiple implementations.

This module is only useful when you know all the implementations ahead
of time. If you want to load arbitrary implementations then you
probably want something like a plugin system, not this module.