Fix build on MirBSD by working around a bug in configure.
The problem is the following: When checking the CFLAGS / LIBS for
including pthread support, any warning on stdout is counted as a no.
However, the test does not use CFLAGS on the command line, which
produces a harmless debug warning. This workaround has the consequence
of adding the default CFLAGS twice on the command line but this is not
a problem.