Update to 1.1.2.
1.1.2:
This bug-fix release corrects two more issues introduced in the xcb port
in the 1.1.0 release: false matching of window names when using -name
and invalid frees when windows have a _NET_WM_NAME property.

It also updates the optional support for the xcb-icccm library to work
with the recent xcb-util-wm 0.3.8 release at the expense of dropping
support for xcb-util 0.3.7 and older.   This support is still only
enabled when specifically requested via configure --with-xcb-icccm.

1.1.1:
This bug-fix release corrects two issues introduced in the xcb port
in the 1.1.0 release: a build issue with the optional support for
the xcb-icccm library, and a runtime issue with reporting coordinates
for windows extending into the negative parts of the coordinate space.

1.1.0:
The major new feature of this release is the rewrite to use libxcb instead
of libX11, batching requests for information to reduce waiting on round-trips
in gathering the client information from the X server.   Testing over a
high-latency connection from California to China and back resulted in an
improvement from 8 minutes to 45 seconds to output the complete window tree
via xwininfo -root -all of a GNOME desktop session with 114 windows.

This release thus libxcb to build, (though it still requires some libX11
headers as well).   It also requires a minimum xproto version of 7.0.17.

Should anyone need to maintain the libX11 version, a 1.0.x branch can be
created in git from the 1.0.5 release base as needed.

This release also adds support for reporting some EMWH window manager hints,
including UTF-8 localized window names from the _NET_WM_NAME property.
(This version does not correctly display names for windows with the WM_NAME
 property set using the COMPOUND_TEXT encoding.   Most current desktops &
 toolkits will set the _NET_WM_NAME, which is used in preference to WM_NAME
 if it is set.)