lang/gcc47: Add support for DragonFly
GCC47 was marked NOT-FOR-DRAGONFLY, so support has been added.

* DragonFly-specific files added via patch mechanism
* Some existing patches modified to add DragonFly configuration items
* dl_iterate_phdr error handling support added (FreeBSD support was altered,
  NetBSD and OpenBSD support is commented out)
* The java language is taken off as a default option

On the i386 platform, the compiler will build from a full bootstrap, but
one of the later stages fails on x86_64.  It fails to find libstdc++.so.6
even though the previous stage library was built and -B, -L flags point
to it.  The cause of the platform-specific build failure isn't clear --
The workaround is to disable the bootstrap on DragonFly so that the compiler
is built in one stage instead of three.  This workaround could have been
limited to the x86_64-DragonFly platform only, but currently is applied to
i386-DragonFly too.