Update to Asterisk 10,8.0.  This release is for fixing AST-2012-012 and AST-2012-013, and some general bugs.
The Asterisk Development Team has announced the release of Asterisk 10.8.0.

The release of Asterisk 10.8.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- AST-2012-012: Resolve AMI User Unauthorized Shell Access through
      ExternalIVR

* --- AST-2012-013: Resolve ACL rules being ignored during calls by
      some IAX2 peers

* --- Handle extremely out of order RFC 2833 DTMF

* --- Resolve severe memory leak in CEL logging modules.

* --- Only re-create an SRTP session when needed

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-10.8.0

Thank you for your continued support of Asterisk!